+++ common/src/main/kotlin/org/valkyrienskies/create_interactive/CreateInteractiveUtil.kt
@@ -5,6 +5,9 @@ import com.simibubi.create.content.contraptions.BlockMovementChecks
 import com.simibubi.create.content.contraptions.Contraption
 import com.simibubi.create.content.contraptions.behaviour.MovementContext
 import com.simibubi.create.content.trains.entity.CarriageContraptionEntity
+import com.simibubi.create.content.trains.entity.TrainRelocator
+import com.simibubi.create.content.trains.track.ITrackBlock
+import com.simibubi.create.content.trains.track.TrackBlock
 import com.simibubi.create.foundation.blockEntity.IMultiBlockEntityContainer
 import net.minecraft.core.BlockPos
 import net.minecraft.nbt.CompoundTag
@@ -32,6 +35,7 @@ import org.valkyrienskies.core.api.ships.properties.ShipTransform
 import org.valkyrienskies.core.apigame.ShipTeleportData
 import org.valkyrienskies.core.apigame.world.properties.DimensionId
 import org.valkyrienskies.core.impl.game.ships.ShipTransformImpl
+import org.valkyrienskies.core.util.expand
 import org.valkyrienskies.create_interactive.mixin.DimensionalCarriageEntityAccessor
 import org.valkyrienskies.create_interactive.mixinducks.AbstractContraptionEntityDuck
 import org.valkyrienskies.create_interactive.mixinducks.ContraptionDuck
@@ -40,12 +44,14 @@ import org.valkyrienskies.create_interactive.mixinducks.OrientedContraptionEntit
 import org.valkyrienskies.mod.common.dimensionId
 import org.valkyrienskies.mod.common.getShipManagingPos
 import org.valkyrienskies.mod.common.shipObjectWorld
+import org.valkyrienskies.mod.common.util.set
 import org.valkyrienskies.mod.common.util.settings
 import org.valkyrienskies.mod.common.util.toBlockPos
 import org.valkyrienskies.mod.common.util.toJOML
 import org.valkyrienskies.mod.common.util.toMinecraft
 import org.valkyrienskies.mod.common.yRange
 import java.lang.ref.WeakReference
+import java.util.Random
 
 object CreateInteractiveUtil {
     fun createShipForContraption(level: ServerLevel, contraption: Contraption, blockPos: BlockPos, blocks: Map<BlockPos, StructureTemplate.StructureBlockInfo> = contraption.blocks): ShipId? {
@@ -93,8 +99,75 @@ object CreateInteractiveUtil {
             // endregion
         }
 
+        if (blocks.isNotEmpty()) {
+            val random = Random()
+            val anchorPos = contraption.anchor
+            // Relocate trains
+            var minPosNotRelative: Vector3i? = null
+            var maxPosNotRelative: Vector3i? = null
+            val posAsJOML = Vector3i()
+            for ((pos, structureInfo) in blocks) {
+                if (structureInfo.state.block is ITrackBlock) {
+                    // Tick the track block to create its track graph immediately (normally create waits until the next tick, but that's too slow for us)
+                    val posInWorld = pos.offset(anchorPos)
+                    val posInShip = pos.offset(shipCenter.toBlockPos())
+                    val stateInWorld = level.getBlockState(posInShip)
+                    if (stateInWorld.block is TrackBlock) {
+                        stateInWorld.block.tick(stateInWorld, level, posInShip, random)
+                    }
+                    posAsJOML.set(posInWorld)
+                    if (minPosNotRelative == null) {
+                        minPosNotRelative = Vector3i(posAsJOML)
+                    } else {
+                        minPosNotRelative.min(posAsJOML)
+                    }
+                    if (maxPosNotRelative == null) {
+                        maxPosNotRelative = Vector3i(posAsJOML)
+                    } else {
+                        maxPosNotRelative.max(posAsJOML)
+                    }
+                }
+            }
+            attemptRelocation(minPosNotRelative, maxPosNotRelative, level, anchorPos, blocks, shipCenter)
+            /*
+            if (minPos != null && maxPos != null) {
+                val searchAABB: AABBdc = AABBd(minPos.x().toDouble(), minPos.y().toDouble(), minPos.z().toDouble(), maxPos.x().toDouble() + 1.0, maxPos.y().toDouble() + 1.0, maxPos.z().toDouble() + 1.0)
+                val trainCars = level.getEntitiesOfClass(CarriageContraptionEntity::class.java, searchAABB.toMinecraft())
+                // Only attempt to relocate the first carriage
+                trainCars.filter { it.carriageIndex == 0 }.forEach { carriageEntity ->
+                    val leadingBogeyPosInLocal: Vector3dc = carriageEntity.anchorVec.toJOML().sub(0.0, 1.0, 0.0)
+                    val closestBlockPos = BlockPos(leadingBogeyPosInLocal.x(), leadingBogeyPosInLocal.y(), leadingBogeyPosInLocal.z())
+                    if (blocks[closestBlockPos]?.state?.block is ITrackBlock) {
+                        // Relocate it!
+                        // TODO: Set the directions properly
+                        TrainRelocator.relocate(carriageEntity.carriage.train, level, closestBlockPos.offset(shipCenter.x(), shipCenter.y(), shipCenter.z()), null, false, Vec3(1.0, 0.0, 0.0), false)
+                    }
+                }
+            }
+
+             */
+        }
+
         serverShip.isStatic = true
         return serverShip.id
+    }
+
+    private fun attemptRelocation(minPosNotRelative: Vector3i?, maxPosNotRelative: Vector3i?, level: Level, offsetPos: BlockPos, localBlocks: Map<BlockPos, StructureTemplate.StructureBlockInfo>, shipCenter: Vector3ic) {
+        if (minPosNotRelative != null && maxPosNotRelative != null) {
+            val searchAABB: AABBdc = AABBd(minPosNotRelative.x().toDouble(), minPosNotRelative.y().toDouble(), minPosNotRelative.z().toDouble(), maxPosNotRelative.x().toDouble() + 1.0, maxPosNotRelative.y().toDouble() + 1.0, maxPosNotRelative.z().toDouble() + 1.0).expand(1.0)
+            val trainCars = level.getEntitiesOfClass(CarriageContraptionEntity::class.java, searchAABB.toMinecraft())
+            // Only attempt to relocate the first carriage
+            // TODO: Filter to only include trains that actually collide with searchAABB
+            trainCars.filter { it.carriageIndex == 0 }.forEach { carriageEntity ->
+                val leadingBogeyPosInLocal: Vector3dc = carriageEntity.anchorVec.toJOML().sub(0.0, 1.0, 0.0)
+                val closestBlockPosRelative = BlockPos(leadingBogeyPosInLocal.x(), leadingBogeyPosInLocal.y(), leadingBogeyPosInLocal.z()).subtract(offsetPos)
+                if (localBlocks[closestBlockPosRelative]?.state?.block is ITrackBlock) {
+                    // Relocate it!
+                    // TODO: Set the directions properly
+                    TrainRelocator.relocate(carriageEntity.carriage.train, level, closestBlockPosRelative.offset(shipCenter.x(), shipCenter.y(), shipCenter.z()), null, false, Vec3(0.0, 0.0, 1.0), false)
+                }
+            }
+        }
     }
 
     fun doesContraptionHaveShipLoaded(contraption: Contraption): Boolean {
@@ -256,7 +329,7 @@ object CreateInteractiveUtil {
 
     fun getShipForMovementContext(context: MovementContext): Ship? = getShipForContraption(context.contraption)
 
-    private fun getShipForContraption(contraption: Contraption): Ship? {
+    internal fun getShipForContraption(contraption: Contraption): Ship? {
         val contraptionEntity = contraption.entity ?: return null
         val shadowShipId = (contraptionEntity as AbstractContraptionEntityDuck).`ci$getShadowShipId`() ?: return null
         return contraptionEntity.level.shipObjectWorld.allShips.getById(shadowShipId)